requirements.txt
kafka-python==2.2.2
pymongo==4.15.3
